Show storage details for all mounted devices
Unsure if it would be preferable to show the filesystem device or the
mount point as the identifier for the device.

I prefer the mount point but others may disagree, it's easy to change it
if required.

I've left the boot partition in for the moment, but again its easy to
exclude this from the display if folks don't think it's useful.